Crate parquet_variant_json

Conversion between JSON and the Variant Binary Encoding from Apache Parquet.

See json_to_variant for converting a JSON string to a Variant.

See variant_to_json for converting a Variant to a JSON string.

🚧 Work In Progress

This crate is under active development and is not yet ready for production use. If you are interested in helping, you can find more information on the GitHub Variant issue

Modules

from_json 🔒

Module for parsing JSON strings as Variant

to_json 🔒

Module for converting Variant data to JSON format

Functions

json_to_variant

Converts a JSON string to Variant to a [VariantBuilderExt], such as VariantBuilder.

variant_to_json

This function writes JSON directly to any type that implements Write, making it efficient for streaming or when you want to control the output destination.

variant_to_json_string

Convert [Variant] to JSON String

variant_to_json_value

Convert [Variant] to serde_json::Value